Add files via upload
This project is about user login and registration. The user can create an account and login an account but there is a problem in the part of login function where you can use others user password ex. let say the first user Username: Jan123 & Password:Frances213  and the second user Username: Jan876 & Password:Frances321. Now if we login if we put Username:Jan123 for the user one and the Password: Frances321 where the second user password it will say true, because I use txt File reading all lines in the files so as long the Password is correspond in every line it will tells true. 

What I want is : if the login is like this I want the result to be false but I still dont have Idea how to be done.